+++ Chess/src/Rook.h
@@ -12,8 +12,8 @@
 class Rook: public Piece {
 public:
 	Rook(bool white, unsigned short x, unsigned short y);
-	void move(int x, int y);
-	bool validMove(int x, int y);
+	void move(unsigned short x, unsigned short y);
+	bool validMove(unsigned short x, unsigned short y);
 };
 
 #endif /* ROOK_H_ */
